python-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
psycopg2-binary
pydantic
requests
apscheduler
google-generativeai
python-dotenv
groq
